+++ packages/app-staking/src/library/Modal/Title.tsx
@@ -4,8 +4,8 @@
 import type { IconProp } from '@fortawesome/fontawesome-svg-core'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { useHelp } from 'hooks/useHelp'
-import { ButtonHelpTooltip } from 'library/ButtonHelpTooltip'
 import type { FunctionComponent, SVGProps } from 'react'
+import { ButtonHelp } from 'ui-buttons'
 import { Header, Title as Wrapper } from 'ui-core/modal'
 import { Close } from 'ui-overlay'
 
@@ -51,7 +51,7 @@ export const Title = ({
 						<Wrapper>
 							{title}
 							{helpKey ? (
-								<ButtonHelpTooltip
+								<ButtonHelp
 									marginLeft
 									definition={helpKey}
 									openHelp={openHelpTooltip}
