+++ packages/app-staking/src/library/StatCards/Timeleft.tsx
@@ -2,8 +2,8 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 import { useHelp } from 'hooks/useHelp'
-import { ButtonHelpTooltip } from 'library/ButtonHelpTooltip'
 import { Countdown } from 'library/Countdown'
+import { ButtonHelp } from 'ui-buttons'
 import { Badge, Stat } from 'ui-core/base'
 import { Pie } from 'ui-graphs'
 import type { TimeleftProps } from './types'
@@ -37,7 +37,7 @@ export const Timeleft = ({
 						<Stat.Subtitle>
 							{label}{' '}
 							{helpKey !== undefined ? (
-								<ButtonHelpTooltip
+								<ButtonHelp
 									marginLeft
 									definition={helpKey}
 									openHelp={openHelpTooltip}
