+++ packages/hooks/src/index.ts
@@ -24,6 +24,7 @@ export * from './usePlugins'
 export * from './usePoolMembers'
 export * from './usePoolSetups'
 export * from './useProxySupported'
+export * from './useShowHelp'
 export * from './useStaking'
 export * from './useStakingAccountCategories'
 export * from './useStakingMetrics'
