+++ packages/ui-buttons/package.json
@@ -20,6 +20,7 @@
 		"@w3ux/react-polkicon": "^4.0.2",
 		"@w3ux/utils": "3.1.1",
 		"classnames": "^2.5.1",
+		"hooks": "workspace:*",
 		"react": "^19.2.7",
 		"ui-styles": "workspace:*",
 		"ui-core": "workspace:*"
